+++ packages/codemirror/rollup.config.ts
@@ -1,0 +1,8 @@
+import withSolid from "rollup-preset-solid";
+
+export default withSolid([
+  {
+    input: "src/index.ts",
+    targets: ["esm", "cjs"],
+  },
+]);