+++ packages/codemirror/src/utils/createExtensions.ts
@@ -2,6 +2,9 @@ import type { Extension } from "@codemirror/state";
 import { createEffect, mergeProps, on } from "solid-js";
 
 export interface ExtensionsProps {
+  /**
+   * An array of CodeMirror extensions to use
+   */
   extensions?: Extension[];
 }