+++ packages/core/test/createCodeMirror.test.tsx
@@ -1,7 +1,8 @@
-import { describe, test, expect, it, vi, afterEach } from "vitest";
+import { describe, expect, it, vi, afterEach } from "vitest";
 import { fireEvent, screen, render } from "solid-testing-library";
 import { createCodeMirror, CodeMirrorProps } from "../src";
 import { createRoot } from "solid-js";
+import { EditorView } from "@codemirror/view";
 
 function CodeMirror(props: CodeMirrorProps) {
   let ref: HTMLDivElement | undefined;
@@ -26,7 +27,17 @@ describe("createCodeMirror", () => {
     });
   });
 
-  it("can be controlled", async () => {
+  it("handles default value", async () => {
+    const { unmount } = render(() => <CodeMirror value="test" />);
+
+    const div = (await screen.findByRole("textbox")) as HTMLDivElement;
+
+    expect(div.textContent).toBe("test");
+
+    unmount();
+  });
+
+  it("can be controlled on value change", async () => {
     const onChangeMock = vi.fn().mockImplementation((code) => {});
 
     const { unmount } = render(() => (
@@ -39,6 +50,19 @@ describe("createCodeMirror", () => {
 
     expect(onChangeMock).toHaveBeenCalledTimes(1);
     expect(onChangeMock).toHaveBeenCalledWith("test");
+
+    unmount();
+  });
+
+  it("can be controlled on editor mount", async () => {
+    const onEditorMountMock = vi.fn().mockImplementation((editor) => {});
+
+    const { unmount } = render(() => (
+      <CodeMirror onEditorMount={onEditorMountMock} />
+    ));
+
+    expect(onEditorMountMock).toHaveBeenCalledTimes(1);
+    expect(onEditorMountMock).toHaveBeenCalledWith(expect.any(EditorView));
 
     unmount();
   });