+++ packages/codemirror/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-codemirror/codemirror
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @solid-codemirror/core@1.0.0
+
 ## 0.0.5
 
 ### Patch Changes